+++ sunpy/physics/tests/test_differential_rotation.py
@@ -15,7 +15,28 @@ from sunpy.time import parse_time
 
 # Please note the numbers in these tests are not checked for physical
 # accuracy, only that they are the values the function was outputting upon
-# implementation.
+# implementation.  This is not a significant issue for the diff_rot function
+# since it is relatively simple and the values it produces can be easily
+# compared to other implementations of the same simple function.  The same
+# cannot be said for the solar_rotate_coordinate function.  This functionality
+# relies accurate knowledge of the solar ephemeris in particular.
+# There is no reference implementation of the solar_rotate_coordinate function
+# of demonstrated trustworthiness at time of writing in any language.  There
+# are no known independent values or tests that can be used to test the
+# veracity of the solar_rotate_coordinate function.  This being the case, the
+# solar_rotate_coordinate function is tested against values that it generated.
+# Therefore these tests test for consistency, not accuracy.  Note that when the
+# 0.8.0 branch was released, the solar ephemeris calculation was handed off to
+# the relevant Astropy code.  The solar_rotate_coordinate tests were changed
+# for self-consistency.  Note that the change in position comparing the results
+# of pre- and 0.8.0 sunpy solar coordinate rotation functionality (rot_hpc
+# and solar_rotate_coordinate respectively) was on the order of 0.5 arcseconds.
+# At time of writing, the difference between the rotation
+# calculated using the pre-0.8.0 rot_hpc function and the SSWIDL equivalent
+# rot_xy.pro for the tests given in pre-0.8.0 were on the order of hundredths
+# of an arcsecond.  I suspect that the reason for the small differences is
+# because the sunpy's ephemeris and coordinate transformation infrastructure
+# was largely based on that in SSWIDL.
 
 
 @pytest.fixture
@@ -76,7 +97,7 @@ def test_solar_rotate_coordinate():
     # Test the coordinate
     np.testing.assert_almost_equal(d.Tx.to(u.arcsec).value, -562.3768, decimal=1)
     np.testing.assert_almost_equal(d.Ty.to(u.arcsec).value, 119.2684, decimal=1)
-    np.testing.assert_almost_equal(d.distance.to(u.km).value, 150083151.9, decimal=1)
+    np.testing.assert_almost_equal(d.distance.to(u.km).value, 150083151.97246578, decimal=1)
 
     # Test that the SkyCoordinate is Helioprojective
     assert isinstance(d.frame, frames.Helioprojective)